requirements.txt
pyspark==3.5.6
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.5
scikit-learn==1.3.2
findspark==2.0.1
tqdm==4.67.1
requests>=2.28.0
mysql-connector-python>=9.0.0
pyarrow>=10.0.0
